Support for saved plots and pre-selected data
Currently plotted data is saved as a json structure to a file when 'Save Plot' is clicked. That file is loaded at runtime to contain a list of pre-saved plots that is used as a menu drop down in the data selection box.
Also cleaned up js file structure, removed initial selected plots.